pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools_scm[toml]>=7.1.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "confidence/_version.py"
local_scheme = "no-local-version"

[tool.setuptools]
packages = {find = {exclude = ["demo"]}}

[project]
name = "spotify_confidence_sdk"
dynamic = ["version"]
description = "Confidence provider for the OpenFeature SDK"
readme = "README.md"
authors = [{ name = "Confidence team" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3"
]
keywords = []
dependencies = [
    "requests==2.32.3", 
    "openfeature-sdk==0.4.2", 
    "typing_extensions==4.9.0",
    "httpx==0.27.2",
    "protobuf==5.29.3"
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest==7.4.2",
    "pytest-mock==3.11.1",
    "black==23.9.1",
    "flake8==6.1.0",
    "requests_mock==1.11.0",
    "setuptools_scm==8.0.4",
    "mypy==1.5.1"
]

[project.urls]
Homepage = "https://github.com/spotify/confidence-sdk-python"
